Replace a couple of long by intptr_t
Replace casting of pointers to long that cannot work on Windows 64 bit (where pointers are 64 bit and long 32 bit)

+++ core/metacling/src/TClingBaseClassInfo.h
@@ -51,8 +51,8 @@ private:
    const clang::Decl            *fDecl; // Current class whose bases we are iterating through, we do *not* own.
    clang::CXXRecordDecl::base_class_const_iterator fIter; // Current iterator.
    TClingClassInfo              *fBaseInfo; // Base class our iterator is currently pointing at, we own.
-   std::vector<std::pair<std::pair<const clang::Decl*, clang::CXXRecordDecl::base_class_const_iterator>, long> > fIterStack; // Iterator stack.
-   long                          fOffset; // Offset of the current base, fDecl, in the most-derived class.
+   std::vector<std::pair<std::pair<const clang::Decl*, clang::CXXRecordDecl::base_class_const_iterator>, intptr_t> > fIterStack; // Iterator stack.
+   intptr_t                      fOffset; // Offset of the current base, fDecl, in the most-derived class.
    bool                          fClassInfoOwnership; // We created the fClassInfo and we need to delete it in the constructor.
 
 public:
@@ -72,7 +72,7 @@ public:
    bool          IsValid() const;
    int           Next();
    int           Next(int onlyDirect);
-   ptrdiff_t     Offset(void * address = 0, bool isDerivedObject = true) const;
+   intptr_t      Offset(void * address = 0, bool isDerivedObject = true) const;
    long          Property() const;
    long          Tagnum() const;
    void          FullName(std::string &output, const ROOT::TMetaUtils::TNormalizedCtxt &normCtxt) const;
